+++ CustomAntBot/MyBot/MyBot/Private/foodseeker.cpp
@@ -1,25 +1,9 @@
-/*
- * Copyright 2011 Nicolai Hähnle
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "foodseeker.h"
+#include "..\Public\foodseeker.h"
 
 #include <algorithm>
 
-#include "astar.h"
-#include "Bot.h"
+#include "..\Public\astar.h"
+#include "..\Public\Bot.h"
 
 using namespace std;
 
@@ -51,13 +35,13 @@ void FoodSeeker::init()
 {
 }
 
-uint FoodSeeker::foodidx_at(const Location & pos)
+size_t FoodSeeker::foodidx_at(const Location & pos)
 {
 	state.bug << "foodidx " << pos << endl;
 
 	//assert(state.grid[pos.row][pos.col].isFood);
 
-	for (uint idx = 0; idx < d.foods.size(); ++idx) {
+	for (size_t idx = 0; idx < d.foods.size(); ++idx) {
 		if (d.foods[idx].where == pos)
 			return idx;
 	}
@@ -71,11 +55,11 @@ uint FoodSeeker::foodidx_at(const Location & pos)
 void FoodSeeker::assign_food()
 {
 	AStar<LocationEvalZero, StepEvalDefault> astar(state, LocationEvalZero(), StepEvalDefault(state));
-	uint unassigned = min(d.foods.size(), bot.m_ants.size());
+	size_t unassigned = min(d.foods.size(), bot.m_ants.size());
 
 	while (unassigned > 0) {
 		state.bug << unassigned << " unassigned" << endl;
-		for (uint foodidx = 0; foodidx < d.foods.size(); ++foodidx) {
+		for (size_t foodidx = 0; foodidx < d.foods.size(); ++foodidx) {
 			if (!d.foods[foodidx].claimed) {
 				astar.push(d.foods[foodidx].where);
 			}
@@ -90,7 +74,7 @@ void FoodSeeker::assign_food()
 			if (state.grid[where.row][where.col].ant != 0)
 				continue;
 
-			uint antidx = bot.myantidx_at(where);
+			size_t antidx = bot.myantidx_at(where);
 			Ant & ant = bot.m_ants[antidx];
 
 			if (ant.assigneddirection)
@@ -119,7 +103,7 @@ void FoodSeeker::run()
 {
 	d.foods.clear();
 	d.foods.reserve(state.food.size());
-	for (uint foodidx = 0; foodidx < state.food.size(); ++foodidx) {
+	for (size_t foodidx = 0; foodidx < state.food.size(); ++foodidx) {
 		d.foods.push_back(Food());
 		Food & food = d.foods.back();
 		food.where = state.food[foodidx];
